+++ CHANGELOG.md
@@ -14,6 +14,12 @@ Types of changes are:
 
 ## [Unreleased]
 
+## [0.5.0] - 2021-11-12
+
+### Changed
+
+- Remove top level heading from the Markdown template.
+
 ### Fixed
 
 - Description of `generate --templates`.
@@ -51,7 +57,8 @@ Add classifiers to the package.
 
 - Initial release
 
-[Unreleased]: https://github.com/radeklat/settings-doc/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/radeklat/settings-doc/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/radeklat/settings-doc/compare/0.5.0...0.5.0
 [0.4.0]: https://github.com/radeklat/settings-doc/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/radeklat/settings-doc/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/radeklat/settings-doc/compare/0.1.1...0.2.0